pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "code_analyzer"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
]